Remove defunct 'start-container.sh' script
dockersh will spin up the container for each user when they
first log in, so we don't need this script any more.

Some of the functionality of this script will still need to be
ported over to dockersh somehow (e.g. setting sysctls, connecting
to the ndag network, etc.). But this is only necessary for making
ndag multicast work; analysis of flowtuple / pcap data from swift
shouldn't need anything special.