Wrap spark-submit in our own script to hide unnecessary detail
The script should simplify the process of submitting spark jobs
for our users, and is installed in /usr/local/sbin/ on the
spark user container.

Added a README with instructions for users to get setup
(i.e. add their EC2 keys to the spark config) and how to submit
a spark job using our script. This is installed into the users
/home/ on the container.

Added a pyspark config file, which can be modified by the user
to change the amount of memory per executor, number of executors
and amount of driver memory. This is also installed automatically
into the container.